requirements.txt
pandas
requests
autogen-agentchat==0.4.3
